+++ src/main/java/net/minecraft/block/BlockCrops.java
@@ -19,7 +19,7 @@ public class BlockCrops extends BlockBush implements IGrowable
 
     protected BlockCrops()
     {
-        this.setDefaultState(this.blockState.getBaseState().withProperty(AGE, Integer.valueOf(0)));
+        this.setDefaultState(this.blockState.getBaseState().withProperty(AGE, 0));
         this.setTickRandomly(true);
         float f = 0.5F;
         this.setBlockBounds(0.5F - f, 0.0F, 0.5F - f, 0.5F + f, 0.25F, 0.5F + f);
@@ -40,7 +40,7 @@ public class BlockCrops extends BlockBush implements IGrowable
 
         if (worldIn.getLightFromNeighbors(pos.up()) >= 9)
         {
-            int i = state.getValue(AGE).intValue();
+            int i = state.getValue(AGE);
 
             if (i < 7)
             {
@@ -48,7 +48,7 @@ public class BlockCrops extends BlockBush implements IGrowable
 
                 if (rand.nextInt((int)(25.0F / f) + 1) == 0)
                 {
-                    worldIn.setBlockState(pos, state.withProperty(AGE, Integer.valueOf(i + 1)), 2);
+                    worldIn.setBlockState(pos, state.withProperty(AGE, i + 1), 2);
                 }
             }
         }
@@ -56,14 +56,14 @@ public class BlockCrops extends BlockBush implements IGrowable
 
     public void grow(World worldIn, BlockPos pos, IBlockState state)
     {
-        int i = state.getValue(AGE).intValue() + MathHelper.getRandomIntegerInRange(worldIn.rand, 2, 5);
+        int i = state.getValue(AGE) + MathHelper.getRandomIntegerInRange(worldIn.rand, 2, 5);
 
         if (i > 7)
         {
             i = 7;
         }
 
-        worldIn.setBlockState(pos, state.withProperty(AGE, Integer.valueOf(i)), 2);
+        worldIn.setBlockState(pos, state.withProperty(AGE, i), 2);
     }
 
     protected static float getGrowthChance(Block blockIn, World worldIn, BlockPos pos)
@@ -82,7 +82,7 @@ public class BlockCrops extends BlockBush implements IGrowable
                 {
                     f1 = 1.0F;
 
-                    if (iblockstate.getValue(BlockFarmland.MOISTURE).intValue() > 0)
+                    if (iblockstate.getValue(BlockFarmland.MOISTURE) > 0)
                     {
                         f1 = 3.0F;
                     }
@@ -142,7 +142,7 @@ public class BlockCrops extends BlockBush implements IGrowable
 
         if (!worldIn.isRemote)
         {
-            int i = state.getValue(AGE).intValue();
+            int i = state.getValue(AGE);
 
             if (i >= 7)
             {
@@ -161,7 +161,7 @@ public class BlockCrops extends BlockBush implements IGrowable
 
     public Item getItemDropped(IBlockState state, Random rand, int fortune)
     {
-        return state.getValue(AGE).intValue() == 7 ? this.getCrop() : this.getSeed();
+        return state.getValue(AGE) == 7 ? this.getCrop() : this.getSeed();
     }
 
     public Item getItem(World worldIn, BlockPos pos)
@@ -171,7 +171,7 @@ public class BlockCrops extends BlockBush implements IGrowable
 
     public boolean canGrow(World worldIn, BlockPos pos, IBlockState state, boolean isClient)
     {
-        return state.getValue(AGE).intValue() < 7;
+        return state.getValue(AGE) < 7;
     }
 
     public boolean canUseBonemeal(World worldIn, Random rand, BlockPos pos, IBlockState state)
@@ -186,12 +186,12 @@ public class BlockCrops extends BlockBush implements IGrowable
 
     public IBlockState getStateFromMeta(int meta)
     {
-        return this.getDefaultState().withProperty(AGE, Integer.valueOf(meta));
+        return this.getDefaultState().withProperty(AGE, meta);
     }
 
     public int getMetaFromState(IBlockState state)
     {
-        return state.getValue(AGE).intValue();
+        return state.getValue(AGE);
     }
 
     protected BlockState createBlockState()
