+++ src/main/java/net/minecraft/block/BlockFlowerPot.java
@@ -31,7 +31,7 @@ public class BlockFlowerPot extends BlockContainer
     public BlockFlowerPot()
     {
         super(Material.circuits);
-        this.setDefaultState(this.blockState.getBaseState().withProperty(CONTENTS, BlockFlowerPot.EnumFlowerType.EMPTY).withProperty(LEGACY_DATA, Integer.valueOf(0)));
+        this.setDefaultState(this.blockState.getBaseState().withProperty(CONTENTS, BlockFlowerPot.EnumFlowerType.EMPTY).withProperty(LEGACY_DATA, 0));
         this.setBlockBoundsForItemRender();
     }
 
@@ -276,7 +276,7 @@ public class BlockFlowerPot extends BlockContainer
 
     public int getMetaFromState(IBlockState state)
     {
-        return state.getValue(LEGACY_DATA).intValue();
+        return state.getValue(LEGACY_DATA);
     }
 
     public IBlockState getActualState(IBlockState state, IBlockAccess worldIn, BlockPos pos)
