+++ src/main/java/net/minecraft/client/renderer/ActiveRenderInfo.java
@@ -70,7 +70,7 @@ public class ActiveRenderInfo
 
             if (iblockstate.getBlock() instanceof BlockLiquid)
             {
-                f = BlockLiquid.getLiquidHeightPercent(iblockstate.getValue(BlockLiquid.LEVEL).intValue()) - 0.11111111F;
+                f = BlockLiquid.getLiquidHeightPercent(iblockstate.getValue(BlockLiquid.LEVEL)) - 0.11111111F;
             }
 
             float f1 = (float)(blockpos.getY() + 1) - f;
