+++ src/main/java/net/minecraft/entity/monster/EntityZombie.java
@@ -91,9 +91,9 @@ public class EntityZombie extends EntityMob
     protected void entityInit()
     {
         super.entityInit();
-        this.getDataWatcher().addObject(12, Byte.valueOf((byte)0));
-        this.getDataWatcher().addObject(13, Byte.valueOf((byte)0));
-        this.getDataWatcher().addObject(14, Byte.valueOf((byte)0));
+        this.getDataWatcher().addObject(12, (byte) 0);
+        this.getDataWatcher().addObject(13, (byte) 0);
+        this.getDataWatcher().addObject(14, (byte) 0);
     }
 
     public int getTotalArmorValue()
@@ -147,7 +147,7 @@ public class EntityZombie extends EntityMob
 
     public void setChild(boolean childZombie)
     {
-        this.getDataWatcher().updateObject(12, Byte.valueOf((byte)(childZombie ? 1 : 0)));
+        this.getDataWatcher().updateObject(12, (byte) (childZombie ? 1 : 0));
 
         if (this.worldObj != null && !this.worldObj.isRemote)
         {
@@ -170,7 +170,7 @@ public class EntityZombie extends EntityMob
 
     public void setVillager(boolean villager)
     {
-        this.getDataWatcher().updateObject(13, Byte.valueOf((byte)(villager ? 1 : 0)));
+        this.getDataWatcher().updateObject(13, (byte) (villager ? 1 : 0));
     }
 
     public void onLivingUpdate()
@@ -567,7 +567,7 @@ public class EntityZombie extends EntityMob
     protected void startConversion(int ticks)
     {
         this.conversionTime = ticks;
-        this.getDataWatcher().updateObject(14, Byte.valueOf((byte)1));
+        this.getDataWatcher().updateObject(14, (byte) 1);
         this.removePotionEffect(Potion.weakness.id);
         this.addPotionEffect(new PotionEffect(Potion.damageBoost.id, ticks, Math.min(this.worldObj.getDifficulty().getDifficultyId() - 1, 0)));
         this.worldObj.setEntityState(this, (byte)16);
