+++ src/main/java/net/minecraft/entity/projectile/EntityArrow.java
@@ -112,7 +112,7 @@ public class EntityArrow extends Entity implements IProjectile
 
     protected void entityInit()
     {
-        this.dataWatcher.addObject(16, Byte.valueOf((byte)0));
+        this.dataWatcher.addObject(16, (byte) 0);
     }
 
     public void setThrowableHeading(double x, double y, double z, float velocity, float inaccuracy)
@@ -530,11 +530,11 @@ public class EntityArrow extends Entity implements IProjectile
 
         if (critical)
         {
-            this.dataWatcher.updateObject(16, Byte.valueOf((byte)(b0 | 1)));
+            this.dataWatcher.updateObject(16, (byte) (b0 | 1));
         }
         else
         {
-            this.dataWatcher.updateObject(16, Byte.valueOf((byte)(b0 & -2)));
+            this.dataWatcher.updateObject(16, (byte) (b0 & -2));
         }
     }
