+++ src/main/java/net/minecraft/item/crafting/FurnaceRecipes.java
@@ -73,7 +73,7 @@ public class FurnaceRecipes
     public void addSmeltingRecipe(ItemStack input, ItemStack stack, float experience)
     {
         this.smeltingList.put(input, stack);
-        this.experienceList.put(stack, Float.valueOf(experience));
+        this.experienceList.put(stack, experience);
     }
 
     public ItemStack getSmeltingResult(ItemStack stack)
@@ -105,7 +105,7 @@ public class FurnaceRecipes
         {
             if (this.compareItemStacks(stack, entry.getKey()))
             {
-                return entry.getValue().floatValue();
+                return entry.getValue();
             }
         }
