+++ src/main/java/net/minecraft/util/ObjectIntIdentityMap.java
@@ -14,7 +14,7 @@ public class ObjectIntIdentityMap<T> implements IObjectIntIterable<T>
 
     public void put(T key, int value)
     {
-        this.identityMap.put(key, Integer.valueOf(value));
+        this.identityMap.put(key, value);
 
         while (this.objectList.size() <= value)
         {
@@ -27,7 +27,7 @@ public class ObjectIntIdentityMap<T> implements IObjectIntIterable<T>
     public int get(T key)
     {
         Integer integer = this.identityMap.get(key);
-        return integer == null ? -1 : integer.intValue();
+        return integer == null ? -1 : integer;
     }
 
     public final T getByValue(int value)
